Laden Speichern Attribute Descriptions

+++ DSAProject/Classes/Charakter/Values/Attribute/BaseAttack.cs
@@ -12,16 +12,17 @@ namespace DSAProject.Classes.Charakter.Values.Attribute
     {
         public BaseAttack(ICharakterAttribut attribute) : base(attribute) { }
         public override string Name => "Attacke-Basis";
-        protected override int Calculate()
+        protected override double Calculate()
         {
             var var1 = Attribute.GetAttributMAXValue(CharakterAttribut.Mut, out Error error);
+                
             Logger.Log(error, nameof(BaseAttack), nameof(Calculate));
             var var2 = Attribute.GetAttributMAXValue(CharakterAttribut.Gewandheit, out error);
             Logger.Log(error, nameof(BaseAttack), nameof(Calculate));
             var var3 = Attribute.GetAttributMAXValue(CharakterAttribut.Körperkraft, out error);
             Logger.Log(error, nameof(BaseAttack), nameof(Calculate));
 
-            return (var1 + var2 + var3) / 5;
+            return (var1 + var2 + var3) / 5.0;
         }
     }
 }
