+++ DSAProject/Classes/Charakter/Values/Attribute/BaseRange.cs
@@ -12,7 +12,7 @@ namespace DSAProject.Classes.Charakter.Values.Attribute
     {
         public BaseRange(ICharakterAttribut attribute) : base(attribute) { }
         public override string Name => "Fernkampf-Basis";
-        protected override int Calculate()
+        protected override double Calculate()
         {
             var var1 = Attribute.GetAttributMAXValue(CharakterAttribut.Intuition, out Error error);
             Logger.Log(error, nameof(BaseRange), nameof(Calculate));
@@ -21,7 +21,7 @@ namespace DSAProject.Classes.Charakter.Values.Attribute
             var var3 = Attribute.GetAttributMAXValue(CharakterAttribut.Körperkraft, out error);
             Logger.Log(error, nameof(BaseRange), nameof(Calculate));
 
-            return (var1 + var2 + var3) / 5;
+            return (var1 + var2 + var3) / 5.0;
         }
     }
 }
