+++ DSAProject/Classes/Charakter/Values/Attribute/WoundSwell.cs
@@ -11,11 +11,11 @@ namespace DSAProject.Classes.Charakter.Values.Attribute
     {
         public WoundSwell(Interfaces.ICharakterAttribut attribute) : base(attribute) { }
         public override string Name => "Wundschwelle";
-        protected override int Calculate()
+        protected override double Calculate()
         {
             var var1 = Attribute.GetAttributMAXValue(CharakterAttribut.Konstitution, out Error error);
             Logger.Log(error, nameof(WoundSwell), nameof(Calculate));
-            return (var1) / 2;
+            return (var1) / 2.0;
         }
     }
 }
